backend/requirements.txt
# Core Dependencies
djangorestframework>=3.14 # For building REST APIs
psycopg2-binary>=2.9      # PostgreSQL database driver
django-cors-headers>=3.14 # To allow React Native to access your API
django-environ>=0.10      # For managing environment variables
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.2.2

